requirements.txt
streamlit
faiss-cpu
sentence-transformers
huggingface-hub